refactor: replace hook with function
