+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/attribute-filter/edit.tsx
@@ -33,12 +33,11 @@ import { EXCLUDED_BLOCKS } from '../../constants';
 import { FilterOptionItem } from '../../types';
 import { InitialDisabled } from '../../components/initial-disabled';
 import { Notice } from '../../components/notice';
-import { useProductFilterClearButtonManager } from '../../hooks/use-product-filter-clear-button-manager';
 
 const ATTRIBUTES = getSetting< AttributeSetting[] >( 'attributes', [] );
 
 const Edit = ( props: EditProps ) => {
-	const { attributes: blockAttributes, clientId } = props;
+	const { attributes: blockAttributes } = props;
 
 	const {
 		attributeId,
@@ -48,7 +47,6 @@ const Edit = ( props: EditProps ) => {
 		showCounts,
 		sortOrder,
 		hideEmpty,
-		clearButton,
 	} = blockAttributes;
 
 	const attributeObject = getAttributeFromId( attributeId );
@@ -58,11 +56,6 @@ const Edit = ( props: EditProps ) => {
 	>( [] );
 	const [ isOptionsLoading, setIsOptionsLoading ] =
 		useState< boolean >( true );
-
-	useProductFilterClearButtonManager( {
-		clientId,
-		showClearButton: clearButton,
-	} );
 
 	const { results: attributeTerms, isLoading: isTermsLoading } =
 		useCollection< AttributeTerm[] >( {
