+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/price-filter/edit.tsx
@@ -19,7 +19,10 @@ import { getAllowedBlocks } from '../../utils';
 import { getPriceFilterData } from './utils';
 import { InitialDisabled } from '../../components/initial-disabled';
 import { BlockAttributes } from './types';
-import { useProductFilterClearButtonManager } from '../../hooks/use-product-filter-clear-button-manager';
+import { toggleProductFilterClearButtonVisibilityFactory } from '../../utils/toggle-product-filter-clear-button-visibility';
+
+const toggleProductFilterClearButtonVisibility =
+	toggleProductFilterClearButtonVisibilityFactory();
 
 const Edit = ( props: BlockEditProps< BlockAttributes > ) => {
 	const { attributes, setAttributes, clientId } = props;
@@ -32,11 +35,6 @@ const Edit = ( props: BlockEditProps< BlockAttributes > ) => {
 		isEditor: true,
 	} );
 
-	useProductFilterClearButtonManager( {
-		clientId,
-		showClearButton: clearButton,
-	} );
-
 	return (
 		<div { ...blockProps }>
 			<InspectorControls group="styles">
@@ -44,9 +42,13 @@ const Edit = ( props: BlockEditProps< BlockAttributes > ) => {
 					<ToggleControl
 						label={ __( 'Clear button', 'woocommerce' ) }
 						checked={ clearButton }
-						onChange={ ( value ) =>
-							setAttributes( { clearButton: value } )
-						}
+						onChange={ ( value ) => {
+							setAttributes( { clearButton: value } );
+							toggleProductFilterClearButtonVisibility( {
+								clientId,
+								showClearButton: value,
+							} );
+						} }
 					/>
 				</PanelBody>
 			</InspectorControls>
