+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/rating-filter/components/inspector.tsx
@@ -17,6 +17,10 @@ import {
  */
 import RatingStars from './rating-stars';
 import type { Attributes } from '../types';
+import { toggleProductFilterClearButtonVisibilityFactory } from '../../../utils/toggle-product-filter-clear-button-visibility';
+
+const toggleProductFilterClearButtonVisibility =
+	toggleProductFilterClearButtonVisibilityFactory();
 
 function MinimumRatingLabel( {
 	stars,
@@ -41,9 +45,13 @@ function MinimumRatingLabel( {
 }
 
 export const Inspector = ( {
+	clientId,
 	attributes,
 	setAttributes,
-}: Pick< BlockEditProps< Attributes >, 'attributes' | 'setAttributes' > ) => {
+}: Pick<
+	BlockEditProps< Attributes >,
+	'attributes' | 'setAttributes' | 'clientId'
+> ) => {
 	const { showCounts, minRating, clearButton } = attributes;
 
 	function setCountVisibility( value: boolean ) {
@@ -125,9 +133,13 @@ export const Inspector = ( {
 					<ToggleControl
 						label={ __( 'Clear button', 'woocommerce' ) }
 						checked={ clearButton }
-						onChange={ ( value ) =>
-							setAttributes( { clearButton: value } )
-						}
+						onChange={ ( value ) => {
+							setAttributes( { clearButton: value } );
+							toggleProductFilterClearButtonVisibility( {
+								clientId,
+								showClearButton: value,
+							} );
+						} }
 					/>
 				</PanelBody>
 			</InspectorControls>
