+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/rating-filter/edit.tsx
@@ -33,10 +33,9 @@ import { Notice } from '../../components/notice';
 import type { Attributes } from './types';
 import './style.scss';
 import { InitialDisabled } from '../../components/initial-disabled';
-import { useProductFilterClearButtonManager } from '../../hooks/use-product-filter-clear-button-manager';
 
 const RatingFilterEdit = ( props: BlockEditProps< Attributes > ) => {
-	const { attributes, setAttributes } = props;
+	const { attributes, setAttributes, clientId } = props;
 
 	const { isPreview, showCounts, minRating } = attributes;
 
@@ -166,11 +165,6 @@ const RatingFilterEdit = ( props: BlockEditProps< Attributes > ) => {
 		minRating,
 	] );
 
-	useProductFilterClearButtonManager( {
-		clientId: props.clientId,
-		showClearButton: attributes.clearButton,
-	} );
-
 	if ( ! filteredCountsLoading && displayedOptions.length === 0 ) {
 		return null;
 	}
@@ -191,6 +185,7 @@ const RatingFilterEdit = ( props: BlockEditProps< Attributes > ) => {
 	return (
 		<>
 			<Inspector
+				clientId={ clientId }
 				attributes={ attributes }
 				setAttributes={ setAttributes }
 			/>
