+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/status-filter/inspector.tsx
@@ -22,8 +22,11 @@ import {
  */
 import { BlockAttributes, EditProps } from './types';
 import { getInnerBlockByName } from '../../utils';
+import { toggleProductFilterClearButtonVisibilityFactory } from '../../utils/toggle-product-filter-clear-button-visibility';
 
 let displayStyleOptions: Block[] = [];
+const toggleProductFilterClearButtonVisibility =
+	toggleProductFilterClearButtonVisibilityFactory();
 
 export const Inspector = ( {
 	attributes,
@@ -119,9 +122,13 @@ export const Inspector = ( {
 					<ToggleControl
 						label={ __( 'Clear button', 'woocommerce' ) }
 						checked={ clearButton }
-						onChange={ ( value ) =>
-							setAttributes( { clearButton: value } )
-						}
+						onChange={ ( value ) => {
+							setAttributes( { clearButton: value } );
+							toggleProductFilterClearButtonVisibility( {
+								clientId,
+								showClearButton: value,
+							} );
+						} }
 					/>
 				</PanelBody>
 			</InspectorControls>
